requirements-dev.txt
# Testing dependencies
pytest
pytest-cov
pytest-asyncio
pytest-homeassistant-custom-component

# Linting dependencies
flake8
black
isort

# Optional: Type checking
mypy